+++ klaw-gui/src/state/workbench.rs
@@ -51,6 +51,14 @@ impl WorkbenchState {
             UiAction::OpenMenu(menu) => self.open_or_activate(menu),
             UiAction::ActivateTab(tab_id) => self.activate(tab_id),
             UiAction::CloseTab(tab_id) => self.close(tab_id),
+            UiAction::CloseWindow
+            | UiAction::ToggleFullscreen
+            | UiAction::MinimizeWindow
+            | UiAction::ZoomWindow
+            | UiAction::StartWindowDrag
+            | UiAction::ShowAbout
+            | UiAction::HideAbout
+            | UiAction::CycleTheme => {}
         }
     }